@@ -1,57 +1,107 @@
 {
-  "name": "JS-Encoder",
-  "version": "3.5.1",
+  "name": "js-encoder",
   "private": true,
-  "author": "Longgererer",
-  "email": "c2b139460q@gmail.com",
-  "homepage": "http://jsencoder.lliiooiill.cn/",
-  "repository": "https://github.com/JS-Encoder/JS-Encoder",
-  "bugs": "https://github.com/JS-Encoder/JS-Encoder/issues",
+  "version": "4.0.0-beta.1",
+  "type": "module",
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build --report"
+    "start": "bash src/scripts/start.sh",
+    "dev": "vite",
+    "build": "vue-tsc --noEmit && vite build",
+    "preview": "vite preview",
+    "lint": "eslint --ext .ts,.vue src/ --fix && stylelint \"./src/**/*.{css,scss,vue}\" --fix",
+    "prepare": "husky install"
+  },
+  "prettier": {
+    "eslintIntegration": true,
+    "semi": false,
+    "singleQuote": false,
+    "jsxSingleQuote": false,
+    "printWidth": 120
   },
   "dependencies": {
-    "@prettier/plugin-pug": "^1.19.1",
-    "@vue/compiler-sfc": "^3.2.29",
-    "acorn": "^8.4.1",
-    "acorn-loose": "^8.1.0",
-    "acorn-walk": "^8.1.1",
-    "axios": "^0.19.2",
-    "codemirror": "^5.62.2",
-    "codemirror-emmet": "^1.0.0",
-    "core-js": "^3.6.5",
-    "csslint": "^1.0.5",
-    "element-ui": "^2.13.2",
-    "estraverse": "^5.2.0",
-    "file-saver": "^2.0.2",
-    "flowchart.js": "^1.15.0",
-    "highlight.js": "^10.1.2",
-    "htmlhint": "^0.14.2",
-    "js-beautify": "^1.13.5",
-    "jshint": "^2.12.0",
-    "jszip": "^3.5.0",
-    "less": "^3.12.2",
-    "marked": "^1.1.1",
-    "prettier": "^2.5.1",
-    "pug": "^3.0.0",
-    "qs": "^6.9.4",
-    "stylus": "^0.54.8",
-    "tern": "^0.24.3",
-    "vue": "^2.6.11",
-    "vue-codemirror": "^4.0.6",
-    "vue-i18n": "^8.18.2",
-    "vuex": "^3.4.0"
+    "@babel/standalone": "^7.24.0",
+    "@codemirror/autocomplete": "^6.11.1",
+    "@codemirror/commands": "^6.3.0",
+    "@codemirror/lang-css": "^6.2.1",
+    "@codemirror/lang-html": "^6.4.6",
+    "@codemirror/lang-javascript": "^6.2.1",
+    "@codemirror/lang-less": "^6.0.1",
+    "@codemirror/lang-markdown": "^6.2.2",
+    "@codemirror/lang-sass": "^6.0.2",
+    "@codemirror/lang-vue": "^0.1.2",
+    "@codemirror/language": "^6.9.2",
+    "@codemirror/legacy-modes": "^6.3.3",
+    "@codemirror/lint": "^6.4.2",
+    "@codemirror/search": "^6.5.4",
+    "@codemirror/state": "^6.3.1",
+    "@codemirror/theme-one-dark": "^6.1.2",
+    "@codemirror/view": "^6.21.4",
+    "@emmetio/codemirror6-plugin": "^0.3.0",
+    "@lezer/highlight": "^1.2.0",
+    "@replit/codemirror-indentation-markers": "^6.5.0",
+    "@replit/codemirror-vscode-keymap": "^6.0.2",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@typescript/vfs": "^1.5.0",
+    "@vue/compiler-sfc": "^3.4.21",
+    "autoprefixer": "^10.4.19",
+    "codemirror": "^6.0.1",
+    "coffeescript": "^2.7.0",
+    "eslint4b-prebuilt": "^6.7.2",
+    "espree": "^10.0.1",
+    "estraverse": "^5.3.0",
+    "hash-sum": "^2.0.0",
+    "highlight.js": "^11.9.0",
+    "htmlhint": "^1.1.4",
+    "jszip": "^3.10.1",
+    "less": "^4.2.0",
+    "marked": "^12.0.1",
+    "marked-highlight": "^2.1.1",
+    "marked-katex-extension": "^5.0.1",
+    "pug": "^3.0.2",
+    "sass": "^1.53.0",
+    "sass-loader": "^13.0.2",
+    "stylelint": "^16.2.0",
+    "stylelint-config-recommended-less": "^3.0.0",
+    "stylelint-config-standard": "^35.0.0",
+    "stylelint-config-standard-scss": "^12.0.0",
+    "stylelint-config-standard-vue": "^1.0.0",
+    "stylus": "^0.63.0",
+    "vite-plugin-eslint": "^1.8.1",
+    "vue": "^3.4.5",
+    "vue-router": "^4.3.2"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-vuex": "~4.5.0",
-    "@vue/cli-service": "~4.5.0",
-    "babel-plugin-component": "^1.1.1",
-    "compression-webpack-plugin": "^6.1.1",
-    "sass": "^1.50.0",
-    "sass-loader": "^10.2.1",
-    "vue-template-compiler": "^2.6.11",
-    "webpack-bundle-analyzer": "^4.4.2"
+    "@commitlint/cli": "^18.4.4",
+    "@commitlint/config-conventional": "^18.4.4",
+    "@types/babel__standalone": "^7.1.7",
+    "@types/coffeescript": "^2.5.7",
+    "@types/espree": "^10.0.0",
+    "@types/estraverse": "^5.1.7",
+    "@types/file-saver": "^2.0.7",
+    "@types/hash-sum": "^1.0.2",
+    "@types/htmlhint": "^1.1.5",
+    "@types/less": "^3.0.6",
+    "@types/node": "^18.0.5",
+    "@types/pug": "^2.0.10",
+    "@types/stylus": "^0.48.42",
+    "@typescript-eslint/eslint-plugin": "^6.13.1",
+    "@typescript-eslint/parser": "^6.13.1",
+    "@vitejs/plugin-vue": "^5.0.2",
+    "eslint": "^8.18.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-vue": "^9.1.1",
+    "husky": "^8.0.3",
+    "pinia": "^2.1.7",
+    "postcss-html": "^1.6.0",
+    "prettier": "^3.2.5",
+    "rollup-plugin-external-globals": "^0.10.0",
+    "rollup-plugin-visualizer": "^5.12.0",
+    "typescript": "^5.4.5",
+    "unplugin-vue-define-options": "^1.4.1",
+    "vite": "^5.0.11",
+    "vite-plugin-compression": "^0.5.1",
+    "vite-plugin-require-transform": "^1.0.21",
+    "vue-tsc": "^1.8.27"
   }
 }
